+++ src/main/java/shelter/animal/dto/request/AnimalPatchRequest.java
@@ -14,7 +14,7 @@ public record AnimalPatchRequest(
         String name,
         AddressPatchRequest address,
         @Positive(message = "A idade precisa ser um valor maior que 0")
-        Double age,
+        Integer age,
         AgeUnit ageUnit,
         @DecimalMin(value = "0.5", message = "O peso não pode ser menor que 0.5kg")
         @DecimalMax(value = "60.0", message = "O peso não pode ser maior que 60kg")
